FOUR-11278:Priority 1 My Task: Enlarge the participant task preview panel
width
